+++ SpeechResponder/CustomFunctions/VoiceDetails.cs
@@ -2,8 +2,7 @@
 using Cottle.Values;
 using EddiSpeechResponder.Service;
 using EddiSpeechService;
-using System.Collections.Generic;
-using Utilities;
+using System.Linq;
 
 namespace EddiSpeechResponder.CustomFunctions
 {
@@ -13,80 +12,43 @@ namespace EddiSpeechResponder.CustomFunctions
         public string name => "VoiceDetails";
         public FunctionCategory Category => FunctionCategory.Details;
         public string description => Properties.CustomFunctions_Untranslated.VoiceDetails;
+
         public NativeFunction function => new NativeFunction((values) =>
         {
             if (values.Count == 0)
             {
-                List<VoiceDetail> voices = new List<VoiceDetail>();
                 if (SpeechService.Instance?.allVoices != null)
                 {
-                    foreach (var vc in SpeechService.Instance.allVoices)
-                    {
-                        if (!vc.name.Contains("Microsoft Server Speech Text to Speech Voice"))
-                        {
-                            voices.Add(new VoiceDetail(
-                                vc.name,
-                                vc.Culture.Parent.EnglishName,
-                                vc.Culture.Parent.NativeName,
-                                vc.Culture.Name,
-                                vc.gender
-                                ));
-                        }
-                    }
+                    return new ReflectionValue(
+                        SpeechService.Instance.allVoices.FirstOrDefault(v =>
+                            v.name == SpeechService.Instance.Configuration.StandardVoice) ?? new object());
                 }
-                return new ReflectionValue(voices);
             }
+
             if (values.Count == 1)
             {
-                VoiceDetail result = null;
-                if (SpeechService.Instance?.allVoices != null && !string.IsNullOrEmpty(values[0].AsString))
+                if (int.TryParse(values[0].AsString, out var seed) && SpeechService.Instance?.allVoices != null)
+                {
+                    var fromSeed = new System.Random(seed);
+                    return new ReflectionValue(SpeechService.Instance.allVoices
+                        .OrderBy(o => fromSeed.Next()).ToList());
+                }
+
+                if (!string.IsNullOrEmpty(values[0].AsString) && SpeechService.Instance?.allVoices != null)
                 {
                     foreach (var vc in SpeechService.Instance.allVoices)
                     {
-                        if (vc.name.ToLowerInvariant().Contains(values[0].AsString.ToLowerInvariant())
-                        && !vc.name.Contains("Microsoft Server Speech Text to Speech Voice"))
+                        if (vc.name.ToLowerInvariant().Contains(values[0].AsString.ToLowerInvariant()))
                         {
-                            result = new VoiceDetail(
-                                vc.name,
-                                vc.Culture.Parent.EnglishName,
-                                vc.Culture.Parent.NativeName,
-                                vc.Culture.Name,
-                                vc.gender
-                                );
-                            break;
+                            return new ReflectionValue(vc);
                         }
                     }
+
+                    return $"Voice \"{values[0].AsString}\" not found.";
                 }
-                return new ReflectionValue(result ?? new object());
             }
+
             return "The VoiceDetails function is used improperly. Please review the documentation for correct usage.";
         }, 0, 1);
-    }
-
-    class VoiceDetail
-    {
-        [PublicAPI]
-        public string name { get; }
-
-        [PublicAPI]
-        public string cultureinvariantname { get; }
-
-        [PublicAPI]
-        public string culturename { get; }
-
-        [PublicAPI]
-        public string culturecode { get; }
-
-        [PublicAPI]
-        public string gender { get; }
-
-        public VoiceDetail(string name, string cultureinvariantname, string culturename, string culturecode, string gender)
-        {
-            this.name = name;
-            this.cultureinvariantname = cultureinvariantname;
-            this.culturename = culturename;
-            this.culturecode = culturecode;
-            this.gender = gender;
-        }
     }
 }